big refactoring - moving messages to dedicated package
