+++ autofit/graphical/declarative/__init__.py
@@ -1,2 +1,1 @@
-from .collection import FactorGraphModel
-from .factor import AnalysisFactor, HierarchicalFactor
+
