--- autofit/graphical/declarative/factor/__init__.py
@@ -1,2 +1,0 @@
-from .analysis import AnalysisFactor
-from .hierarchical import HierarchicalFactor
