+++ autofit/graphical/factor_graphs/graph.py
@@ -10,8 +10,9 @@ import numpy as np
 from autofit.graphical.factor_graphs.abstract import FactorValue, AbstractNode
 from autofit.graphical.factor_graphs.factor import Factor
 from autofit.graphical.utils import (
-    add_arrays, aggregate, Axis, cached_property, rescale_to_artists
+    add_arrays, aggregate, Axis, rescale_to_artists
 )
+from autofit.tools.cached_property import cached_property
 from autofit.mapper.variable import Variable, Plate
