+++ autofit/mapper/prior/prior.py
@@ -1,13 +1,10 @@
 import inspect
-import math
 import random
 import sys
 from abc import ABC, abstractmethod
 from typing import Union, Tuple
 
 import numpy as np
-from scipy import stats
-from scipy.special import erfcinv
 
 from autoconf import conf
 from autofit import exc
@@ -169,81 +166,6 @@ class Prior(Variable, ABC, ArithmeticMixin):
     @property
     def limits(self) -> Tuple[float, float]:
         return self.lower_limit, self.upper_limit
-
-
-class GaussianPrior(Prior):
-    """A prior with a gaussian distribution"""
-
-    __name__ = "gaussian_prior"
-
-    def __init__(self, mean, sigma, lower_limit=-math.inf, upper_limit=math.inf):
-        super().__init__(lower_limit, upper_limit)
-        self.mean = float(mean)
-        self.sigma = float(sigma)
-
-        self._log_pdf = None
-
-    @property
-    def logpdf(self):
-        if self._log_pdf is None:
-            norm = stats.norm(
-                loc=self.mean,
-                scale=self.sigma
-            )
-            self._log_pdf = norm.logpdf
-        return self._log_pdf
-
-    def __call__(self, x):
-        return self.logpdf(x)
-
-    def value_for(self, unit):
-        """
-
-        Parameters
-        ----------
-        unit: Float
-            A unit hypercube value between 0 and 1
-        Returns
-        -------
-        value: Float
-            A value for the attribute biased to the gaussian distribution
-        """
-        return self.mean + (self.sigma * math.sqrt(2) * erfcinv(2.0 * (1.0 - unit)))
-
-    def log_prior_from_value(self, value):
-        """
-        Returns the log prior of a physical value, so the log likelihood of a model evaluation can be converted to a
-        posterior as log_prior + log_likelihood.
-
-        This is used by Emcee in the log likelihood function evaluation.
-
-        Parameters
-        ----------
-        value : float
-            The physical value of this prior's corresponding parameter in a `NonLinearSearch` sample."""
-        return (value - self.mean) ** 2.0 / (2 * self.sigma ** 2.0)
-
-    def __str__(self):
-        """The line of text describing this prior for the model_mapper.info file"""
-        return (
-            f"GaussianPrior, mean = {self.mean}, sigma = {self.sigma}"
-        )
-
-    def __repr__(self):
-        return (
-            "<GaussianPrior id={} mean={} sigma={} "
-            "lower_limit={} upper_limit={}>".format(
-                self.id, self.mean, self.sigma, self.lower_limit, self.upper_limit
-            )
-        )
-
-    @property
-    def dict(self) -> dict:
-        """
-        A dictionary representation of this prior
-        """
-        prior_dict = super().dict
-        return {**prior_dict, "mean": self.mean, "sigma": self.sigma}
 
 
 class UniformPrior(Prior):
