+++ autofit/messages/beta.py
@@ -1,10 +1,62 @@
+import warnings
 from typing import Union
 
 import numpy as np
-from scipy.special import betaln, psi
+from scipy.special import betaln
+from scipy.special import psi, polygamma
 
+from autofit.tools.cached_property import cached_property
 from ..messages.abstract import AbstractMessage
-from ..utils import cached_property, inv_beta_suffstats
+
+
+def grad_betaln(ab):
+    psiab = psi(ab.sum(axis=1, keepdims=True))
+    return psi(ab) - psiab
+
+
+def jac_grad_betaln(ab):
+    psi1ab = polygamma(1, ab.sum(axis=1, keepdims=True))
+    fii = polygamma(1, ab) - psi1ab
+    fij = -psi1ab[:, 0]
+    return np.array([[fii[:, 0], fij], [fij, fii[:, 1]]]).T
+
+
+def inv_beta_suffstats(lnX, ln1X):
+    """Solve for a, b for,
+
+    psi(a) + psi(a + b) = lnX
+    psi(b) + psi(a + b) = ln1X
+    """
+    _lnX, _ln1X = np.ravel(lnX), np.ravel(ln1X)
+    lnXs = np.c_[_lnX, _ln1X]
+
+    # Find initial starting location
+    Gs = np.exp(lnXs)
+    dG = 1 - Gs.sum(axis=1, keepdims=True)
+    ab = np.maximum(1, (1 + Gs / dG) / 2)
+
+    # 5 Newton Raphson itertions is generally enough
+    for i in range(5):
+        f = grad_betaln(ab) - lnXs
+        jac = jac_grad_betaln(ab)
+        ab += np.linalg.solve(jac, - f)
+
+    if np.any(ab < 0):
+        warnings.warn(
+            "invalid negative parameters found for inv_beta_suffstats, "
+            "clampling value to 0.5",
+            RuntimeWarning
+        )
+        b = np.clip(ab, 0.5, None)
+
+    shape = np.shape(lnX)
+    if shape:
+        a = ab[:, 0].reshape(shape)
+        b = ab[:, 1].reshape(shape)
+    else:
+        a, b = ab[0, :]
+
+    return a, b
 
 
 class BetaMessage(AbstractMessage):
